refactor(consortium-manual): move logic from endpoint to plugin class
