+++ packages/cactus-test-plugin-consortium-manual/src/test/typescript/integration/plugin-consortium-manual/get-consortium-jws-endpoint.test.ts
@@ -22,7 +22,7 @@ import {
   ConsortiumDatabase,
   ConsortiumMember,
 } from "@hyperledger/cactus-core-api";
-import { PluginRegistry } from "@hyperledger/cactus-core";
+import { PluginRegistry, ConsortiumRepository } from "@hyperledger/cactus-core";
 
 test("member node public keys and hosts are pre-shared", async (t: Test) => {
   const consortiumId = uuidV4();
@@ -138,6 +138,7 @@ test("member node public keys and hosts are pre-shared", async (t: Test) => {
     pluginInstance: [],
   };
 
+  const consortiumRepo = new ConsortiumRepository({ db: consortiumDatabase });
   t.comment(`Setting up first node...`);
   {
     // 2. Instantiate plugin registry which will provide the web service plugin with the key value storage plugin
@@ -150,6 +151,7 @@ test("member node public keys and hosts are pre-shared", async (t: Test) => {
       keyPairPem: keyPair1.toPEM(true),
       consortiumDatabase,
       logLevel: "trace",
+      consortiumRepo,
     };
     const pluginConsortiumManual = new PluginConsortiumManual(options);
 
@@ -204,6 +206,7 @@ test("member node public keys and hosts are pre-shared", async (t: Test) => {
       keyPairPem: keyPair2.toPEM(true),
       consortiumDatabase,
       logLevel: "trace",
+      consortiumRepo,
     };
     const pluginConsortiumManual = new PluginConsortiumManual(options);
 
@@ -259,6 +262,7 @@ test("member node public keys and hosts are pre-shared", async (t: Test) => {
       keyPairPem: keyPair3.toPEM(true),
       consortiumDatabase,
       logLevel: "trace",
+      consortiumRepo,
     };
     const pluginConsortiumManual = new PluginConsortiumManual(options);
