+++ packages/cactus-test-plugin-consortium-manual/src/test/typescript/integration/plugin-consortium-manual/security-isolation-via-api-server-ports.ts
@@ -15,7 +15,7 @@ import {
   Consortium,
   ConsortiumDatabase,
 } from "@hyperledger/cactus-core-api";
-import { PluginRegistry } from "@hyperledger/cactus-core";
+import { PluginRegistry, ConsortiumRepository } from "@hyperledger/cactus-core";
 
 import { PluginKeychainMemory } from "@hyperledger/cactus-plugin-keychain-memory";
 import {
@@ -83,6 +83,8 @@ tap.test(
       consortium: [consortium],
     };
 
+    const consortiumRepo = new ConsortiumRepository({ db: consortiumDatabase });
+
     // 3. Instantiate the web service consortium plugin which will host itself on a new TCP port for isolation/security
     // Note that if we omitted the `webAppOptions` object that the web service plugin would default to installing itself
     // on the default port of the API server. This allows for flexibility in deployments.
@@ -96,6 +98,7 @@ tap.test(
         hostname: "127.0.0.1",
         port: 0,
       },
+      consortiumRepo,
     };
     const pluginConsortiumManual = new PluginConsortiumManual(options);
